Change variable scope calculation to use Labels
Instead of relying on line numbers, rely on Labels to determine
the correct scope of a variable.

This solves problems with generated source code that has no
line number annotations in the bytecode.
It also solves problems with weird formatted tests, where the
scope can not be obviously determined by the line numbers.

Fix for issues #19 and #46