API to configure instrumentation (#54)
Add config API to allow fine-grained inclusion and exclusion of classes and methods:

- include and exclude packages and classes
- include and exclude methods and classes based on their annotations
- exclude methods by name
- include lambdas
- include lambdas only when at least one method is instrumented in the class
- exclude methods shorter than a given minimum loc

The API also capable to enable individual instrumentation features:

- track method start
- track returns
- track unhandled exceptions
- track local variable state changes
- track field state changes
- inject JUnit4 Rule or JUnit5 Extension required for test reporting

Note that Scott's configuration possibilities and default behavior does not change.
These parameters are only available in the internal API to allow other modules to depend on Scott.